+++ app/PageBuilder/Gutenberg/Blocks/Archive.php
@@ -141,6 +141,11 @@ class Archive extends Blocks_Basis {
 			}
 		}
 
+		// mark "showFilter" as deprecated.
+		if( ! empty( $attributes['showFilter'] ) ) {
+			_deprecated_argument( 'PersonioIntegrationLight:Archive:render', 4.0, __( 'Please your the filter blocks instead of the filter on archive block!', 'personio-integration-light' ) );
+		}
+
 		// collect all settings for this block.
 		$attribute_defaults = array(
 			'templates'         => $templates,